feat: implement database schema and models for products and sales tracking

+++ app/Models/Product.php
@@ -52,7 +52,7 @@ class Product extends Model
             'is_wholeprice' => 'boolean',
             'buy_price' => 'decimal:2',
             'sell_price' => 'decimal:2',
-            'stock' => 'integer',
+            'stock' => 'float',
         ];
     }
